feat: モジュールおよび機能管理の実装
モジュール関連のリクエストを処理する ModuleController を追加

ApplicationFeature、ModuleDef、ModuleFeatures のDTOを作成

ApplicationFeature、ModuleDefinition、ModuleFeatureRelation のエンティティクラスを導入

ApplicationFeature および ModuleDefinition 用のフォームを作成

各エンティティに対するJPAリポジトリを作成

機能およびモジュールのCRUD操作を行うサービスを実装

既存の機能・モジュールのマッパーを新しいエンティティ構成に合わせてリファクタリング

旧来の機能・モジュール関連のクラスとサービスを削除

コンテキストサービスを拡張し、機能・モジュールのデータ取得に対応

XMLマッパーファイルを新しいエンティティパスと構造に更新

+++ src/main/java/com/example/projectmanagement/persistence/modules/designs/app/features/FeatureMapper.java
@@ -4,20 +4,20 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import com.example.projectmanagement.modules.designs.app.features.datastructures.entity.ApplicationFeatureEntity;
-import com.example.projectmanagement.modules.designs.app.features.datastructures.entity.ModuleFeatureRelationEntity;
+import com.example.projectmanagement.modules.designs.app.datastructures.entity.ApplicationFeatureEntity;
+import com.example.projectmanagement.modules.designs.app.datastructures.entity.ModuleFeatureRelationEntity;
 
 @Mapper
 public interface FeatureMapper {
 
 	public ApplicationFeatureEntity findById();
-	public List<ApplicationFeatureEntity> findByIdList(List<Integer> list);
+	public List<ApplicationFeatureEntity> findByIdList(List<Long> list);
 	public void insertApplicationFeature(ApplicationFeatureEntity domain);
 	public void updateApplicationFeature(ApplicationFeatureEntity domain);
-	public void deleteApplicationFeature(int targetId);
+	public void deleteApplicationFeature(Long targetId);
 	
 	public void insertFeatureModuleRelation(ModuleFeatureRelationEntity rel);
 	
-	public List<ModuleFeatureRelationEntity> findModuleFeatureByModuleIdList(List<Integer> list);
-	public List<ModuleFeatureRelationEntity> findModuleFeatureByFeatureIdList(List<Integer> list);
+	public List<ModuleFeatureRelationEntity> findModuleFeatureByModuleIdList(List<Long> list);
+	public List<ModuleFeatureRelationEntity> findModuleFeatureByFeatureIdList(List<Long> list);
 }
